OK, khtml_part.lo compiles as well
svn path=/branches/work/kde4/kdelibs/; revision=414317

+++ khtml/khtml_part.cpp
@@ -6060,7 +6060,7 @@ void KHTMLPart::khtmlMouseDoubleClickEvent( khtml::MouseDoubleClickEvent *event 
   }
 }
 
-void KHTMLPart::extendSelection( DOM::NodeImpl* node, long offset, DOM::Node& selectionNode, long& selectionOffset, bool right, bool selectLines )
+void KHTMLPart::extendSelection( DOM::NodeImpl* node, int offset, DOM::Node& selectionNode, long& selectionOffset, bool right, bool selectLines )
 {
   khtml::RenderObject* obj = node->renderer();
